requirements.txt
asgiref==3.11.1
brotli==1.2.0
cffi==2.0.0
cssselect2==0.9.0
Django==5.2.12
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.62.1
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
pydyf==0.12.1
PyJWT==2.12.0
pyphen==0.17.2
python-decouple==3.8
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.1.0
tzdata==2025.3
weasyprint==68.1
webencodings==0.5.1
zopfli==0.4.1
